--- a/Gemfile
+++ b/Gemfile
@@ -12,6 +12,7 @@ gem 'federal_register', '~> 0.7.6'
 gem 'saxerator'
 gem 'pg_search'
 gem 'kaminari'
+gem 'scenic'
 
 # Use Puma as the app server
 gem 'puma', '~> 4.1'