--- a/db/migrate/20201230230531_create_sorn_searches.rb
+++ b/db/migrate/20201230230531_create_sorn_searches.rb
@@ -1,0 +1,7 @@
+class CreateSornSearches < ActiveRecord::Migration[6.0]
+  def change
+    create_view :sorn_searches, materialized: true
+
+    add_index :sorn_searches, :tsv_document, using: :gin
+  end
+end