This PR is an experiment in database tuning. It adds:
- single materialized view with tsvectors for all columns at once.
- an indexes for this view
It will need some new code if people want to search for certain columns, not everything.

#### Results
For single word searches its about three times faster.
For multi word searches, its one second faster.


http://localhost:3000/ with 5890 SORNs -
1. blank search  `Completed 200 OK in 126ms (Views: 62.3ms | ActiveRecord: 63.0ms | Allocations: 35705)`
2. search for ssn `Completed 200 OK in 234ms (Views: 96.6ms | ActiveRecord: 119.2ms | Allocations: 77922)`
3. multiword search for social security number `Completed 200 OK in 3081ms (Views: 50.5ms | ActiveRecord: 1619.7ms | Allocations: 1081335)`

**main branch results**
1. blank search - `Completed 200 OK in 123ms (Views: 57.9ms | ActiveRecord: 64.3ms | Allocations: 35703)`
2. search for ssn `Completed 200 OK in 2907ms (Views: 82.8ms | ActiveRecord: 2813.2ms | Allocations: 76911)`
3. multiword search social security number `Completed 200 OK in 3632ms (Views: 121.9ms | ActiveRecord: 2464.4ms | Allocations: 685153)`